get SNS controller to modern ACK runtime
Hooooooo boy, this was a rough one. :)

We had not updated the ACK runtime or aws-sdk-go in 2 *years* for the
SNS controller, which means that a whole bunch of features in the code
generator and runtime were not able to be activated for SNS.

This PR gets the SNS controller regenerated with the latest ACK runtime
and aws-sdk-go based on a code-generator checkout that contains a bunch
of fixes for attribute-based APIs.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>