get SNS controller to modern ACK runtime (#9)
Hooooooo boy, this was a rough one. :)

We had not updated the ACK runtime or aws-sdk-go in 2 *years* for the SNS controller, which means that a whole bunch of features in the code generator and runtime were not able to be activated for SNS.

This PR gets the SNS controller regenerated with the latest ACK runtime and aws-sdk-go based on a code-generator checkout ([PR here](https://github.com/aws-controllers-k8s/code-generator/pull/376)) that contains a bunch of fixes for attribute-based APIs.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.