updating the olm config file with proper resources (#10)
Issue #, if available:
N/A
Description of changes:
Updating the OLM config file to include the current resources. Also generated the OpenShift bundle and put this in a cluster to make sure the controller started properly. See screen shot below.

![image](https://user-images.githubusercontent.com/88156657/207965630-14f98bf5-a403-4650-8908-98384abc275e.png)

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.

Signed-off-by: Adam D. Cornett <adc@redhat.com>