support Subscription resources
This patch adds support for a Subscription resource in the SNS API. The
CREATE call for Subscription is `Subscribe`, with the DELETE call being
`Unsubscribe`. There are `GetSubscriptionAttributes` and
`SetSubscriptionAttributes` API calls for GET_ATTRIBUTES and
SET_ATTRIBUTES respectively, though this patch does not implement the
UPDATE code paths, leaving it as an empty function that will return an
error  for now.

The Subscription resource has a number of usage caveats in order to make
it fit into the Kubernetes declarative resource model and behaviour
expectations. Notably, we constrain Subscribe operations to only topics
and endpoints that are owned by the same AWS account as the subscriber
and where the protocol is NOT "http(s)" or "email". This is because
there is a manual 2-step multi-account ConfirmSubscription workflow that
is required if the endpoint is owned by a different AWS account or those
protocols are used.

Currently the update code path is not working but the
CREATE/DELETE/READ_ONE code paths are all working.

Tags are not yet supported.

Issue aws-controllers-k8s/community#1491

Signed-off-by: Jay Pipes <jaypipes@gmail.com>